Update the cost module
