+++ src/openscm_calibration/cost/__init__.py
@@ -1,0 +1,3 @@
+"""
+Cost calculations
+"""
